GP-26351 Fix precedence for total_amount and explicitly set payment instrument
This changes the precedence for total_amount. The new order is:
- Template Contribution
- Recurring Contribution

Amounts used by any other previous contributions are ignored.

This also adds the payment instrument as an explicit parameter in
a few API calls in order to avoid core using the instrument that's
set for the payment processor. Core assumes a 1:1 map between
payment processors and instruments, which is not very practical.